linkalong-backend/requirements.txt
amqp==2.6.0
asgiref==3.2.10
attrs==19.3.0
autobahn==20.6.2
Automat==20.2.0
billiard==3.6.3.0
boto==2.49.0
boto3==1.14.20
botocore==1.17.20
celery==4.4.6
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
Django==3.0.8
django-cors-headers==3.4.0
djangorestframework==3.11.0
docutils==0.15.2
future==0.18.2
gensim==3.8.3
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
jmespath==0.10.0
joblib==0.16.0
kombu==4.6.11
lazysorted==0.1.1
model-mommy==2.0.0
nltk==3.5
numpy==1.19.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
regex==2020.6.8
requests==2.24.0
s3transfer==0.3.3
scipy==1.5.1
service-identity==18.1.0
six==1.15.0
smart-open==2.1.0
sqlparse==0.3.1
tqdm==4.47.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
vine==1.3.0
zope.interface==5.1.0
